reti-webapp/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
drf-yasg==1.21.7
django-cors-headers==4.2.0
django-filter==23.3
django-auth-ldap==4.1.0
python-ldap==3.4.3 